+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to the full browser extension will be documented in this fil
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Fixed
+- Initialize the Zest script also when injecting the content script to record the page accessed after starting the recording, as opposed to starting after accessing the page.
+
 ## 0.0.13 - 2025-05-09
 
 ### Added